# docs/swagger.yaml
basePath: /
definitions:
...
  http.CreateRouteRequest:
    properties:
      capacity:
        type: integer
      description:
        type: string
      end_station:
        type: string
      end_time:
        type: string
      interval_minutes:
        type: integer
      name:
        type: string
      start_station:
        type: string
      start_time:
        type: string
    type: object
  http.ListRoutesResponse:
    properties:
      routes:
        items:
          $ref: '#/definitions/models.RouteTransfer'
        type: array
      ticket_draft_id:
        type: integer
    type: object
  http.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  http.SignUpRequest:
    properties:
      full_name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  http.UpdateRouteRequest:
    properties:
      capacity:
        type: integer
      description:
        type: string
      end_station:
        type: string
      end_time:
        type: string
      interval_minutes:
        type: integer
      name:
        type: string
      start_station:
        type: string
      start_time:
        type: string
    type: object
  models.RouteTransfer:
    properties:
      capacity:
        type: integer
      description:
        type: string
      end_station:
        type: string
      end_time:
        type: string
      id:
        type: integer
      image_uuid:
        type: string
      interval_minutes:
        type: integer
      name:
        type: string
      start_station:
        type: string
      start_time:
        type: string
    type: object
...
  models.UserTransfer:
    properties:
      full_name:
        type: string
      id:
        type: integer
      is_moderator:
        type: boolean
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: yarik_tri@mail.ru
    name: Rip API
  description: Rip
  title: Rip
  version: 1.0.1
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Create session
      parameters:
      - description: Username and password
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/http.LoginRequest'
      responses:
        "200":
          description: User logined
        "400":
          description: Incorrect input
          schema: {}
        "500":
          description: Server error
          schema: {}
      summary: Login
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout
      responses:
        "200":
          description: User logined
        "400":
          description: Incorrect input
          schema: {}
        "500":
          description: Server error
          schema: {}
      summary: Logout
      tags:
      - Auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Create account
      parameters:
      - description: User info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/http.SignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Created user
          schema:
            $ref: '#/definitions/models.UserTransfer'
        "400":
          description: Incorrect input
          schema: {}
        "500":
          description: Server error
          schema: {}
      summary: Sign Up
      tags:
      - Auth
  /routes:
    get:
      description: Get all routes
      produces:
      - application/json
      responses:
        "200":
          description: Got routes
          schema:
            $ref: '#/definitions/http.ListRoutesResponse'
        "400":
          description: Incorrect input
          schema: {}
        "500":
          description: Server error
          schema: {}
      summary: List routes
      tags:
      - Routes
    post:
      consumes:
      - application/json
      description: Create route
      parameters:
      - description: Route info
        in: body
        name: routeInfo
        required: true
        schema:
          $ref: '#/definitions/http.CreateRouteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Route created
          schema:
            $ref: '#/definitions/models.RouteTransfer'
        "400":
          description: Incorrect input
          schema: {}
        "500":
          description: Server error
          schema: {}
      summary: Create route
      tags:
      - Routes
  /routes/{routeID}:
    delete:
      description: Delete route by ID
      parameters:
      - description: Route ID
        in: path
        name: routeID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Route deleted
        "400":
          description: Incorrect input
          schema: {}
        "500":
          description: Server error
          schema: {}
      summary: Delete route
      tags:
      - Routes
    get:
      description: Get route by ID
      parameters:
      - description: Route ID
        in: path
        name: routeID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Got route
          schema:
            $ref: '#/definitions/models.RouteTransfer'
        "400":
          description: Incorrect input
          schema: {}
        "500":
          description: Server error
          schema: {}
      summary: Get route
      tags:
      - Routes
    put:
      consumes:
      - application/json
      description: Update route by ID
      parameters:
      - description: Route ID
        in: path
        name: routeID
        required: true
        type: integer
      - description: Route info
        in: body
        name: routeInfo
        required: true
        schema:
          $ref: '#/definitions/http.UpdateRouteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated route
          schema:
            $ref: '#/definitions/models.RouteTransfer'
        "400":
          description: Incorrect input
          schema: {}
        "500":
          description: Server error
          schema: {}
      summary: Update route
      tags:
      - Routes
  /routes/{routeID}/image:
    put:
      consumes:
      - multipart/form-data
      description: Put image of route by ID
      parameters:
      - description: Route image
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Route image updated
        "400":
          description: Incorrect input
          schema: {}
        "500":
          description: Server error
          schema: {}
      summary: Put route image
      tags:
      - Routes
...
schemes:
- https
- http
swagger: "2.0"